+++ src/app/api/whitelist/[id]/route.tsx
@@ -2,8 +2,15 @@ import connectMongoDB from "@/lib/mongodb";
 import Whitelist from "@/models/whitelist";
 import { NextResponse } from "next/server";
 import { ZodError } from "zod";
+import { headers } from "next/headers";
 
 export async function DELETE({ params: { id } }: { params: { id: string } }) {
+  const discordKey = headers().get("discord-key");
+
+  //Allow access only to the Discord Bot
+  if (!discordKey || discordKey !== process.env.BOT_KEY)
+    return new Response("401");
+
   try {
     await connectMongoDB();
     await Whitelist.findOneAndDelete({ idDiscord: id });
